Moved all code into website app
