--- yw_website/yw_db/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class YwDbConfig(AppConfig):
-    name = 'yw_db'
