--- yw_website/yw_rest_services/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class YwRestServicesConfig(AppConfig):
-    name = 'yw_rest_services'
